Move the tour related utils one level up
