+++ plugins/woocommerce-admin/client/guided-tours/add-product-tour/use-track-publish-button.ts
@@ -7,7 +7,7 @@ import { recordEvent } from '@woocommerce/tracks';
 /**
  * Internal dependencies
  */
-import { bindPublishClickEvent } from './utils';
+import { bindPublishClickEvent } from '../utils';
 
 export const useTrackPublishButton = ( showTour: boolean ) => {
 	const unbindPublishClickEvent = useRef< () => void >( () => {} );
