+++ plugins/woocommerce-admin/client/guided-tours/utils.ts
